fix: use sharedarraybuffer
